Bluetooth: Audio: Implement bt_cap_initiator_broadcast_audio_start
Implement the CAP initiator broadcast audio start procedure.

This basically just verifies that the stream context
is set in the metadata, and then calls the BAP
procedure.

We define a new opaque struct, bt_cap_broadcast_source,
that is just an abstraction of the bt_audio_broadcast_source
but is used to indicate that these functions require a
broadcast source created by the CAP function (and not the
BAP) function.

+++ include/zephyr/bluetooth/audio/lc3.h
@@ -316,8 +316,8 @@ enum bt_codec_config_type {
 #define BT_CODEC_LC3_CONFIG_META(_stream_context) \
 { \
 	BT_CODEC_DATA(BT_AUDIO_METADATA_TYPE_STREAM_CONTEXT, \
-		      _stream_context, \
-		      _stream_context >> 8), \
+		      (uint8_t)_stream_context, \
+		      (uint8_t)_stream_context >> 8), \
 }
 
 /** @def BT_CODEC_LC3_CONFIG
